+++ Assets/Scripts/PlaRangeProjectile.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a19ec8ec5208a354a89c0cdf46cfd5ae
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
